Fit Poisson GLM-PCA Model to Count Data

Fit a Poisson GLM-PCA model by maximum-likelihood.

fit_glmpca_pois(
  Y,
  K,
  fit0 = init_glmpca_pois(Y, K),
  tol = 0.0001,
  min_iter = 1,
  max_iter = 100,
  verbose = TRUE,
  control = list()
)

fit_glmpca_pois_control_default()

init_glmpca_pois(
  Y,
  K,
  U,
  V,
  X = numeric(0),
  Z = numeric(0),
  B = numeric(0),
  W = numeric(0),
  fixed_b_cols = numeric(0),
  fixed_w_cols = numeric(0),
  col_size_factor = TRUE,
  row_intercept = TRUE
)

Arguments

Y

The n x m matrix of counts; all entries of Y should be non-negative. It can be a sparse matrix (class "dgCMatrix") or dense matrix (class "matrix").

K

Integer 1 or greater specifying the rank of the matrix factorization. This should only be provided if the initial fit (fit0) is not.

fit0

Initial model fit. It should be an object of class “glmpca_fit_pois”, such as an output from init_glmpca_pois or a previous call to fit_glmpca_pois.

tol

The optimization stops when the change in the log-likelihood between two successive iterations is less than this amount.

min_iter

Minimum number of updates to be performed.

max_iter

Maximum number of updates to be performed.

verbose

If verbose = TRUE, information about the algorithm's progress is printed after each update.

control

List of control parameters to modify behavior of the optimization algorithm; see “Details”.

U

An optional argument giving the initial estimate of the loadings matrix. It should be an n x K matrix, where n is the number of rows in the counts matrix Y, and K > 0 is the rank of the matrix factorization. When U and V are not provided, input argument K should be specified instead.

V

An optional argument giving is the initial estimate of the factors matrix. It should be a m x K matrix, where m is the number of columns in the counts matrix Y, and K > 0 is the rank of the matrix factorization. When U and V are not provided, input argument K should be specified instead.

X

Optional argument giving row covariates of the count matrix Y. It should be an n x nx matrix, where nx is the number of row covariates.

Z

Optional argument giving column covariates of the count matrix Y. It should be an m x nz matrix, where nz is the number of column covariates.

B

Optional argument giving the initial estimates for the coefficients of the row covariates. It should be an m x nx matrix, where nx is the number of row covariates. This argument is ignored if X is not provided.

W

Optional argument giving the initial estimates for the coefficients of the column covariates. It should be an n x nz matrix, where nz is the number of column covariates. This argument is ignored if Z is not provided.

fixed_b_cols

Optional numeric vector specifying which columns of B (if any) should be fixed during optimization. This argument is ignored if X is not provided.

fixed_w_cols

Optional numeric vector specifying which columns of W (if any) should be fixed during optimization. This argument is ignored if Z is not provided.

col_size_factor

If col_size_factor = TRUE, add a fixed factor accounting for average differences in Poisson rates across columns of Y. Setting col_size_factor = TRUE and row_intercept = TRUE is intended to replicate the default behavior of glmpca.

row_intercept

If row_intercept = TRUE, add a fixed factor accounting for average differences in Poisson rates across rows of Y. Setting col_size_factor = TRUE and row_intercept = TRUE is intended to replicate the default behavior of glmpca.

Value

An object capturing the state of the model fit. It contains estimates of \(U\), \(V\) and \(D\) (stored as matrices

U, V and a vector of diagonal entries d, analogous to the svd return value); the other parameters (\(X\), \(B\), \(Z\), \(W\)); the log-likelihood achieved (loglik); information about which columns of

\(B\) and \(W\) are fixed (fixed_b_cols,

fixed_w_cols); and a data frame progress storing information about the algorithm's progress after each update.

Details

In generalized principal component analysis (GLM-PCA) based on a Poisson likelihood, the counts \(y_{ij}\) stored in an \(n \times m\) matrix \(Y\) are modeled as $$y_{ij} \sim Pois(\lambda_{ij}),$$ in which the logarithm of each rate parameter \(\lambda_{ij}\) is defined as a linear combination of rank-K matrices to be estimated from the data: $$\log \lambda_{ij} = (UDV')_{ij},$$ where \(U\) and \(V\) are orthogonal matrices of dimension \(n \times K\) and \(m \times K\), respectively, and \(D\) is a diagonal \(K \times K\) matrix in which the entries along its diagonal are positive and decreasing. \(K\) is a tuning parameter specifying the rank of the matrix factorization. This is the same as the low-rank matrix decomposition underlying PCA (that is, the singular value decomposition), but because we are not using a linear (Gaussian) model, this is called “generalized PCA” or “GLM PCA”.

To allow for additional components that may be fixed, fit_glmpca_pois can also fit the more general model $$\log \lambda_{ij} = (UDV' + XB' + WZ')_{ij},$$ in which \(X\), \(Z\) are fixed matrices of dimension \(n \times n_x\) and \(m \times n_z\), respectively, and \(B\), \(W\) are matrices of dimension \(m \times n_x\) and \(n \times n_z\) to be estimated from the data.

fit_glmpca_pois computes maximum-likelihood estimates (MLEs) of \(U\), \(V\), \(D\), \(B\) and \(W\) satistifying the orthogonality constraints for \(U\) and \(V\) and the additional constraints on \(D\) that the entries are positive and decreasing. This is accomplished by iteratively fitting a series of Poisson GLMs, where each of these individual Poissons GLMs is fitted using a fast cyclic co-ordinate descent (CCD) algorithm.

The control argument is a list in which any of the following named components will override the default optimization algorithm settings (as they are defined by fit_glmpca_pois_control_default):

num_ccd_iter

Number of co-ordinate descent updates to be made to parameters at each iteration of the algorithm.

line_search

If line_search = TRUE, a backtracking line search is performed at each iteration of CCD to guarantee improvement in the objective (the log-likelihood).

alpha

alpha parameter for backtracking line search. (Should be a number between 0 and 0.5, typically a number near zero.)

beta

beta parameter for backtracking line search controlling the rate at which the step size is decreased. (Should be a number between 0 and 0.5.)

calc_deriv

If calc_deriv = TRUE, the maximum gradient of \(U\) and \(V\) is calculated and stored after each update. This may be useful for assessing convergence of the optimization, though increases overhead.

calc_max_diff

If calc_max_diff = TRUE, the largest change in \(U\) and \(V\) after each update is calculated and stored. This may be useful for monitoring progress of the optimization algorithm.

orthonormalize

If orthonormalize = TRUE, the matrices \(U\) and \(V\) are made to be orthogonal after each update step. This generally improves the speed of convergence while incurring minimal overhead.

References

Townes, F. W., Hicks, S. C., Aryee, M. J. and Irizarry, R. A. (2019). Feature selection and dimension reduction for single-cell RNA-Seq based on a multinomial model. Genome Biology 20, 295. https://doi.org/10.1186/s13059-019-1861-6

Collins, M., Dasgupta, S. and Schapire, R. E. (2002). A generalization of principal components analysis to the exponential family. In Advances in Neural Information Processing Systems 14.

See also

Examples

set.seed(1)
n <- 1000
p <- 500
K <- 3
dat  <- generate_glmpca_data_pois(n,p,K)
fit0 <- init_glmpca_pois(dat$Y,K)
fit  <- fit_glmpca_pois(dat$Y,fit0 = fit0)
#> Fitting GLM-PCA model to 1000 x 500 count matrix.
#> Iteration 1: log-likelihood = -7.545705756902e+05
#> Iteration 2: log-likelihood = -5.881488243774e+05
#> Iteration 3: log-likelihood = -5.750941070175e+05
#> Iteration 4: log-likelihood = -5.674175424878e+05
#> Iteration 5: log-likelihood = -5.632980720695e+05
#> Iteration 6: log-likelihood = -5.612572786791e+05
#> Iteration 7: log-likelihood = -5.602690873177e+05
#> Iteration 8: log-likelihood = -5.597851880160e+05
#> Iteration 9: log-likelihood = -5.595411110457e+05
#> Iteration 10: log-likelihood = -5.594128823879e+05
#> Iteration 11: log-likelihood = -5.593421361361e+05
#> Iteration 12: log-likelihood = -5.593008809284e+05
#> Iteration 13: log-likelihood = -5.592753590856e+05
#> Iteration 14: log-likelihood = -5.592586189133e+05
#> Iteration 15: log-likelihood = -5.592470378957e+05
#> Iteration 16: log-likelihood = -5.592386606051e+05
#> Iteration 17: log-likelihood = -5.592323870206e+05
#> Iteration 18: log-likelihood = -5.592275680357e+05
#> Iteration 19: log-likelihood = -5.592237998686e+05
#> Iteration 20: log-likelihood = -5.592208173963e+05
#> Iteration 21: log-likelihood = -5.592184375099e+05
#> Iteration 22: log-likelihood = -5.592165281303e+05
#> Iteration 23: log-likelihood = -5.592149906656e+05
#> Iteration 24: log-likelihood = -5.592137496191e+05
#> Iteration 25: log-likelihood = -5.592127461263e+05
#> Iteration 26: log-likelihood = -5.592119337166e+05
#> Iteration 27: log-likelihood = -5.592112753987e+05
#> Iteration 28: log-likelihood = -5.592107415588e+05
#> Iteration 29: log-likelihood = -5.592103084010e+05
#> Iteration 30: log-likelihood = -5.592099567568e+05
#> Iteration 31: log-likelihood = -5.592096711562e+05
#> Iteration 32: log-likelihood = -5.592094390981e+05
#> Iteration 33: log-likelihood = -5.592092504708e+05
#> Iteration 34: log-likelihood = -5.592090970898e+05
#> Iteration 35: log-likelihood = -5.592089723254e+05
#> Iteration 36: log-likelihood = -5.592088708044e+05
#> Iteration 37: log-likelihood = -5.592087881696e+05
#> Iteration 38: log-likelihood = -5.592087208871e+05
#> Iteration 39: log-likelihood = -5.592086660882e+05
#> Iteration 40: log-likelihood = -5.592086214439e+05
#> Iteration 41: log-likelihood = -5.592085850624e+05
#> Iteration 42: log-likelihood = -5.592085554066e+05
#> Iteration 43: log-likelihood = -5.592085312269e+05
#> Iteration 44: log-likelihood = -5.592085115073e+05
#> Iteration 45: log-likelihood = -5.592084954213e+05
#> Iteration 46: log-likelihood = -5.592084822964e+05
#> Iteration 47: log-likelihood = -5.592084715853e+05
#> Iteration 48: log-likelihood = -5.592084628423e+05
#> Iteration 49: log-likelihood = -5.592084557043e+05
#> Iteration 50: log-likelihood = -5.592084498756e+05
#> Iteration 51: log-likelihood = -5.592084451152e+05
#> Iteration 52: log-likelihood = -5.592084412266e+05
#> Iteration 53: log-likelihood = -5.592084380497e+05
#> Iteration 54: log-likelihood = -5.592084354537e+05
#> Iteration 55: log-likelihood = -5.592084333320e+05
#> Iteration 56: log-likelihood = -5.592084315978e+05
#> Iteration 57: log-likelihood = -5.592084301801e+05
#> Iteration 58: log-likelihood = -5.592084290209e+05
#> Iteration 59: log-likelihood = -5.592084280731e+05
#> Iteration 60: log-likelihood = -5.592084272980e+05
#> Iteration 61: log-likelihood = -5.592084266639e+05
#> Iteration 62: log-likelihood = -5.592084261454e+05
#> Iteration 63: log-likelihood = -5.592084257214e+05
#> Iteration 64: log-likelihood = -5.592084253745e+05
#> Iteration 65: log-likelihood = -5.592084250912e+05
#> Iteration 66: log-likelihood = -5.592084248594e+05
#> Iteration 67: log-likelihood = -5.592084246697e+05
#> Iteration 68: log-likelihood = -5.592084245146e+05
#> Iteration 69: log-likelihood = -5.592084243879e+05
#> Iteration 70: log-likelihood = -5.592084242845e+05
#> Iteration 71: log-likelihood = -5.592084241996e+05